+++ tests/testthat/test_detect_changepoints.R
@@ -5,7 +5,6 @@ source("objects_for_tests.R")
 
 test_that("safely_detect_changepoints", {
   expect_equal(safely_detect_changepoints(c(2,2,2,2,6,6,6,6,6)), 4)
-  expect_is(safely_detect_changepoints(c(2,2,2,2,6,6,6,6,6)), "integer")
   expect_equal(safely_detect_changepoints(data = NULL), NULL)
 })